update FAN to 4.20 Android and 4.22 iOS
Using this on an iOS build may require you to update the `ld` linker in your Adobe Air SDK distribution at `{AIR SDK directory}/lib/aot/bin/ld64/` by copying /usr/bin/ld` to that directory and renaming it `ld64`, overwriting the one there. It seems that the linker provided with Adobe AIR is outdated, and the old one gives errors about the FAN SDK not having slices for arm64 and armv7, even though it does.